+++ connectors/hadoop/src/main/java/io/unitycatalog/hadoop/internal/UCHadoopConf.java
@@ -1,4 +1,4 @@
-package io.unitycatalog.hadoop;
+package io.unitycatalog.hadoop.internal;
 
 import io.unitycatalog.client.retry.JitterDelayRetryPolicy;
 import io.unitycatalog.client.retry.RetryPolicy;
