+++ connectors/hadoop/src/main/java/io/unitycatalog/hadoop/internal/auth/CredPropsUtil.java
@@ -1,8 +1,8 @@
-package io.unitycatalog.hadoop.auth;
+package io.unitycatalog.hadoop.internal.auth;
 
-import static io.unitycatalog.hadoop.UCHadoopConf.FS_AZURE_ACCOUNT_AUTH_TYPE_PROPERTY_NAME;
-import static io.unitycatalog.hadoop.UCHadoopConf.FS_AZURE_ACCOUNT_IS_HNS_ENABLED;
-import static io.unitycatalog.hadoop.UCHadoopConf.FS_AZURE_SAS_TOKEN_PROVIDER_TYPE;
+import static io.unitycatalog.hadoop.internal.UCHadoopConf.FS_AZURE_ACCOUNT_AUTH_TYPE_PROPERTY_NAME;
+import static io.unitycatalog.hadoop.internal.UCHadoopConf.FS_AZURE_ACCOUNT_IS_HNS_ENABLED;
+import static io.unitycatalog.hadoop.internal.UCHadoopConf.FS_AZURE_SAS_TOKEN_PROVIDER_TYPE;
 
 import io.unitycatalog.client.auth.TokenProvider;
 import io.unitycatalog.client.model.AwsCredentials;
@@ -11,7 +11,7 @@ import io.unitycatalog.client.model.GcpOauthToken;
 import io.unitycatalog.client.model.PathOperation;
 import io.unitycatalog.client.model.TableOperation;
 import io.unitycatalog.client.model.TemporaryCredentials;
-import io.unitycatalog.hadoop.UCHadoopConf;
+import io.unitycatalog.hadoop.internal.UCHadoopConf;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
