+++ connectors/hadoop/src/test/java/io/unitycatalog/hadoop/internal/auth/CredPropsUtilTest.java
@@ -1,4 +1,4 @@
-package io.unitycatalog.hadoop.auth;
+package io.unitycatalog.hadoop.internal.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
