+++ connectors/spark/src/main/java/io/unitycatalog/spark/fs/CredScopedKey.java
@@ -1,6 +1,6 @@
 package io.unitycatalog.spark.fs;
 
-import io.unitycatalog.hadoop.UCHadoopConf;
+import io.unitycatalog.hadoop.internal.UCHadoopConf;
 import java.net.URI;
 import java.util.Objects;
 import org.apache.hadoop.conf.Configuration;
