+++ connectors/spark/src/test/java/io/unitycatalog/spark/auth/storage/BaseCredRenewITTest.java
@@ -7,7 +7,7 @@ import io.delta.tables.DeltaTable;
 import io.unitycatalog.client.internal.Clock;
 import io.unitycatalog.client.model.CreateCatalog;
 import io.unitycatalog.client.model.CreateSchema;
-import io.unitycatalog.hadoop.UCHadoopConf;
+import io.unitycatalog.hadoop.internal.UCHadoopConf;
 import io.unitycatalog.server.base.BaseCRUDTest;
 import io.unitycatalog.server.base.ServerConfig;
 import io.unitycatalog.server.base.catalog.CatalogOperations;
