+++ connectors/spark/src/test/java/io/unitycatalog/spark/fs/CredScopedKeyTest.java
@@ -2,7 +2,7 @@ package io.unitycatalog.spark.fs;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.unitycatalog.hadoop.UCHadoopConf;
+import io.unitycatalog.hadoop.internal.UCHadoopConf;
 import java.net.URI;
 import org.apache.hadoop.conf.Configuration;
 import org.junit.jupiter.api.Test;
